Replace strncpy() with memcpy() in pslr_scsi_linux.c.
Compilers like recent GCC versions will complain about using strncpy() in pslr_scsi_linux.c. This is because strncpy() is traditionally used with length argument dependent on destination string length. In this code, however, the length is dependent on source string size. It is therefore recommended to use memcpy() instead.